% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_item.R
\name{download_item}
\alias{download_item}
\title{Download NHS TRUD item}
\usage{
download_item(item, directory = ".", download_file = "archive", release = NULL)
}
\arguments{
\item{item}{An integer, the item to be downloaded.}

\item{directory}{Path to the directory to which this item will be downloaded
to. This is set to the current working directory by default.}

\item{download_file}{The item file to be downloaded. Valid values:
\itemize{
\item \code{"archive"} (the release item)
\item \code{"checksum"}
\item \code{"signature"}
\item \code{"publickKey"}
}}

\item{release}{The name of a specific release ID to be downloaded (this can
be ascertained using \code{\link[=get_item_metadata]{get_item_metadata()}}). If \code{NULL} (default), then the
latest item release will be downloaded.}
}
\value{
The file path to the downloaded file, returned invisibly.
}
\description{
Downloads files for a specified NHS TRUD item. By default this downloads
the latest release.

\emph{\strong{Subscription Required}}

You must subscribe to TRUD items individually through the NHS TRUD website
before you can access them using \code{\link[=get_item_metadata]{get_item_metadata()}} or \code{\link[=download_item]{download_item()}}.
Simply having an API key is not sufficient. To see items you're already
subscribed to, use \code{\link[=get_subscribed_metadata]{get_subscribed_metadata()}}. To browse all available
items, use \code{\link[=trud_items]{trud_items()}}.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Download Community Services Data Set pre-deadline extract XML Schema
x <- download_item(394, directory = tempdir())

# List downloaded files
unzip(x, list = TRUE)

# Download a previous release
release <- get_item_metadata(394)$releases[[2]]$id

y <- download_item(394, directory = tempdir(), release = release)

unzip(y, list = TRUE)
\dontshow{\}) # examplesIf}
# An informative error is raised if your API key is invalid or missing
try(withr::with_envvar(c("TRUD_API_KEY" = ""), download_item(394)))
}
