% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trud-package.R
\docType{package}
\name{trud-package}
\alias{trud}
\alias{trud-package}
\title{trud: Query the NHS TRUD API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A convenient R interface to the NHS TRUD API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/trud/}
  \item \url{https://github.com/rmgpanw/trud}
  \item Report bugs at \url{https://github.com/rmgpanw/trud/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

}
\keyword{internal}
