% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trud-package.R
\docType{package}
\name{trud-package}
\alias{trud}
\alias{trud-package}
\title{trud: Query the 'NHS TRUD API'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a convenient R interface to the 'National Health Service NHS Technology Reference Update Distribution (TRUD) API', allowing users to list available releases for their subscribed items, retrieve metadata, and download release files. For more information on the API, see \url{https://isd.digital.nhs.uk/trud/users/guest/filters/0/api}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/trud/}
  \item \url{https://github.com/ropensci/trud}
  \item Report bugs at \url{https://github.com/ropensci/trud/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick.19@ucl.ac.uk} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Robert Luben (\href{https://orcid.org/0000-0002-5088-6343}{ORCID})
  \item Abraham Olvera-Barrios (\href{https://orcid.org/0000-0002-3305-4465}{ORCID})
  \item Chuin Ying Ung (\href{https://orcid.org/0000-0001-8487-4589}{ORCID})
}

}
\keyword{internal}
