% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trud_items.R
\name{trud_items}
\alias{trud_items}
\title{Get available NHS TRUD items}
\usage{
trud_items()
}
\value{
A tibble, with columns \code{item_number} and \code{item_name}.
}
\description{
Scrapes \href{https://isd.digital.nhs.uk/trud/users/guest/filters/0/categories/1}{this page}
from the NHS TRUD website for all available items.
}
\examples{
\dontrun{
 trud_items()
}
}
